requirements.txt
aliyun-python-sdk-core==2.11.4
aliyun-python-sdk-core-v3==2.11.4
aliyun-python-sdk-kms==2.5.0
certifi==2018.11.29
chardet==3.0.4
crcmod==1.7
Django==1.11.11
djangorestframework==3.9.0
drf-extensions==0.4.0
idna==2.8
oss2==2.6.0
pika==0.12.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycryptodome==3.7.2
PyJWT==1.7.1
pytz==2018.7
requests==2.21.0
shortuuid==0.5.0
urllib3==1.24.1
